move error hook to hooks file
