+++ frontend/app/components/auth/auth.utils.ts
@@ -1,8 +1,4 @@
-import { useMemo, useState } from 'preact/hooks';
-import { useIntl } from 'react-intl';
-
 import { isJwtExpired } from 'utils/jwt';
-import { errorMessages, RequestError } from 'utils/errorUtils';
 import { StaticStore } from 'common/static-store';
 import type { FormProvider, OAuthProvider } from 'common/types';
 
@@ -30,39 +26,4 @@ export function getTokenInvalidReason(token: string): null | keyof typeof messag
   }
 
   return null;
-}
-
-export function useErrorMessage(): [string | null, (e: unknown) => void] {
-  const intl = useIntl();
-  const [invalidReason, setInvalidReason] = useState<string | null>(null);
-
-  return useMemo(() => {
-    let errorMessage = invalidReason;
-
-    if (invalidReason && messages[invalidReason]) {
-      errorMessage = intl.formatMessage(messages[invalidReason]);
-    }
-
-    if (invalidReason && errorMessages[invalidReason]) {
-      errorMessage = intl.formatMessage(errorMessages[invalidReason]);
-    }
-
-    function setError(err: unknown): void {
-      if (err === null) {
-        setInvalidReason(null);
-        return;
-      }
-
-      if (typeof err === 'string') {
-        setInvalidReason(err);
-        return;
-      }
-
-      const errorReason = err instanceof RequestError ? err.error : err instanceof Error ? err.message : 'error.0';
-
-      setInvalidReason(errorReason);
-    }
-
-    return [errorMessage, setError];
-  }, [intl, invalidReason]);
 }
